[refactor] Race 책임 분리
- InputView와 OutputView로 사용자 입출력 역할을 나눔

- InputView는 사용자 입력을 담당하는 책임

- OutputView는 사용자 출력을 담당하는 책임

- Race에서 직접 하던 작업을 나눠주면서 분리
