+++ src/main/java/racingcar/View/InputView.java
@@ -1,0 +1,9 @@
+package racingcar.View;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class InputView {
+    public static String getUserInput() {
+        return Console.readLine();
+    }
+}
